backend/bot/serv/requirements.txt
boto3==1.40.17
colorlog==6.9.0
flask==3.0.3
Flask-Cors==5.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.29
langchain-anthropic==0.3.13
langchain-aws==0.2.19
pymongo==4.6.0
python-dotenv==1.0.1
python-multipart==0.0.20
uvicorn==0.34.0
Unidecode==1.3.8
